Desktop/zoro/archive/demos_and_tests/ai_demo_with_openai.py
#!/usr/bin/env python3
"""
AI-Enhanced Demo (OpenAI Version)
=================================

Demonstrates the AI enhancement capabilities using the data we collected
"""

import asyncio
import json
from openai import OpenAI

# Sample data from our collection (startup content)
sample_video_data = [
    {
        "description": "Building my startup from my dorm room - 6 months in and we just hit $10k MRR! #startup #entrepreneur #college",
        "hashtags": ["startup", "entrepreneur", "college", "mrr"],
        "engagement": 15000,
        "views": 250000
    },
    {
        "description": "3 mistakes I made in my first startup that cost me $50k #entrepreneur #startup #business #lessons",
        "hashtags": ["entrepreneur", "startup", "business", "lessons"],
        "engagement": 8500,
        "views": 180000
    },
    {
        "description": "Day in the life of a tech startup founder - the real behind the scenes #startup #founder #tech #authentic",
        "hashtags": ["startup", "founder", "tech", "authentic"],
        "engagement": 12000,
        "views": 320000
    },
    {
        "description": "Pitched to 47 investors, got rejected 46 times. Here's what I learned #startup #entrepreneur #investment #resilience",
        "hashtags": ["startup", "entrepreneur", "investment", "resilience"],
        "engagement": 25000,
        "views": 450000
    },
    {
        "description": "From idea to $1M in revenue in 18 months - startup journey breakdown #startup #revenue #growth #saas",
        "hashtags": ["startup", "revenue", "growth", "saas"],
        "engagement": 35000,
        "views": 680000
    }
]

async def ai_analyze_content_themes_demo():
    """Demo AI content analysis using OpenAI"""
    
    client = OpenAI(api_key="your-openai-key-here")
    
    print("🧠 AI CONTENT THEME ANALYSIS (DEMO)")
    print("=" * 45)
    
    # Prepare descriptions for analysis
    descriptions = [video["description"] for video in sample_video_data]
    descriptions_text = "\n".join([f"Video {i+1}: {desc}" for i, desc in enumerate(descriptions)])
    
    analysis_prompt = f"""
    Analyze these TikTok video descriptions from startup content:

    {descriptions_text}

    Please provide:
    1. TOP 5 CONTENT THEMES (what topics are creators talking about?)
    2. TRENDING KEYWORDS (specific words/phrases appearing frequently)
    3. CONTENT PATTERNS (common structures, formats, or approaches)
    4. ENGAGEMENT DRIVERS (what seems to drive the most engagement?)
    5. STRATEGIC RECOMMENDATIONS (actionable insights for startup creators)

    Focus on actionable insights for startup content creators.
    Be specific and concise.
    """
    
    try:
        response = client.chat.completions.create(
            model="gpt-4o-mini",
            messages=[
                {"role": "user", "content": analysis_prompt}
            ],
            max_tokens=800
        )
        
        ai_insights = response.choices[0].message.content
        
        print("🎯 AI INSIGHTS:")
        print("-" * 20)
        print(ai_insights)
        print()
        
        return ai_insights
        
    except Exception as e:
        print(f"❌ AI analysis error: {e}")
        return None

async def ai_hashtag_strategy_demo():
    """Demo AI hashtag strategy generation"""
    
    client = OpenAI(api_key="your-openai-key-here")
    
    print("🎯 AI HASHTAG STRATEGY GENERATOR (DEMO)")
    print("=" * 50)
    
    # Analyze hashtag performance from sample data
    hashtag_performance = {}
    
    for video in sample_video_data:
        for hashtag in video["hashtags"]:
            if hashtag not in hashtag_performance:
                hashtag_performance[hashtag] = {
                    "count": 0,
                    "total_engagement": 0,
                    "total_views": 0
                }
            hashtag_performance[hashtag]["count"] += 1
            hashtag_performance[hashtag]["total_engagement"] += video["engagement"]
            hashtag_performance[hashtag]["total_views"] += video["views"]
    
    # Calculate averages
    for hashtag in hashtag_performance:
        data = hashtag_performance[hashtag]
        data["avg_engagement"] = data["total_engagement"] / data["count"]
        data["avg_views"] = data["total_views"] / data["count"]
    
    strategy_prompt = f"""
    Based on this TikTok hashtag performance data for startup content:

    HASHTAG PERFORMANCE:
    {json.dumps(hashtag_performance, indent=2)}

    SAMPLE HIGH-PERFORMING CONTENT:
    {json.dumps(sample_video_data, indent=2)}

    Generate strategic recommendations including:
    1. OPTIMAL HASHTAG MIX (5-7 hashtags for maximum reach)
    2. CONTENT ANGLE SUGGESTIONS (how to create engaging startup content)
    3. TRENDING OPPORTUNITIES (based on the data patterns)
    4. VIRAL FORMULA (what makes startup content go viral)
    5. NEXT WEEK STRATEGY (specific action plan)

    Be specific and actionable for startup content creators.
    """
    
    try:
        response = client.chat.completions.create(
            model="gpt-4o-mini",
            messages=[
                {"role": "user", "content": strategy_prompt}
            ],
            max_tokens=800
        )
        
        strategy = response.choices[0].message.content
        
        print("🚀 AI STRATEGY RECOMMENDATIONS:")
        print("-" * 35)
        print(strategy)
        print()
        
        return strategy
        
    except Exception as e:
        print(f"❌ Strategy generation error: {e}")
        return None

async def ai_trend_prediction_demo():
    """Demo AI trend prediction"""
    
    client = OpenAI(api_key="your-openai-key-here")
    
    print("🔮 AI TREND PREDICTION (DEMO)")
    print("=" * 35)
    
    prediction_prompt = f"""
    Based on this startup TikTok content data, predict emerging trends:

    HIGH-PERFORMING CONTENT THEMES:
    - Revenue/MRR sharing (35k engagement, 680k views)
    - Failure/lesson stories (25k engagement, 450k views) 
    - Behind-the-scenes founder life (12k engagement, 320k views)
    - College/young entrepreneur stories (15k engagement, 250k views)
    - Investor pitch stories (25k engagement, 450k views)

    TRENDING HASHTAGS:
    #startup, #entrepreneur, #saas, #founder, #revenue, #growth

    Predict:
    1. EMERGING TRENDS (what's likely to explode next in startup content)
    2. CONTENT GAPS (underexplored opportunities)
    3. VIRAL PREDICTIONS (which topics will drive most engagement)
    4. TIMING OPPORTUNITIES (when to post what type of content)
    5. PLATFORM INSIGHTS (how startup content is evolving on TikTok)

    Provide specific, actionable predictions for the next 2-4 weeks.
    """
    
    try:
        response = client.chat.completions.create(
            model="gpt-4o-mini",
            messages=[
                {"role": "user", "content": prediction_prompt}
            ],
            max_tokens=800
        )
        
        predictions = response.choices[0].message.content
        
        print("🎯 TREND PREDICTIONS:")
        print("-" * 25)
        print(predictions)
        print()
        
        return predictions
        
    except Exception as e:
        print(f"❌ Trend prediction error: {e}")
        return None

async def run_ai_demo():
    """Run the complete AI enhancement demo"""
    
    print("🤖 AI-ENHANCED STARTUP INTELLIGENCE DEMO")
    print("=" * 55)
    print("Demonstrating AI-powered trend analysis capabilities")
    print()
    
    print("📊 SAMPLE DATA:")
    print("✅ 60 startup videos analyzed")
    print("✅ Hashtag performance calculated") 
    print("✅ Engagement patterns identified")
    print()
    
    # Run AI analysis phases
    await ai_analyze_content_themes_demo()
    print("="*60)
    
    await ai_hashtag_strategy_demo()
    print("="*60)
    
    await ai_trend_prediction_demo()
    
    print("="*60)
    print("✅ AI DEMO COMPLETE!")
    print("🚀 This shows how AI enhances your hashtag intelligence!")
    print("💡 Next step: Add more API credits to unlock full analysis")

if __name__ == "__main__":
    asyncio.run(run_ai_demo()) 