Desktop/zoro/requirements.txt
# TikTok Trend Prediction Multi-Agent System Requirements

# LangChain & Multi-Agent Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.7
langgraph==0.0.16
langsmith==0.0.75

# LLM Providers
langchain-openai==0.0.5
langchain-anthropic==0.0.4

# LlamaIndex & RAG Framework
llama-index==0.10.0
llama-index-core==0.10.0
llama-index-llms-openai==0.1.5
llama-index-llms-anthropic==0.3.5
llama-index-embeddings-openai==0.1.5
llama-index-vector-stores-chroma==0.1.3
chromadb==0.4.18

# Database & Data Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pandas==2.1.4
duckdb==0.9.2

# API & HTTP Requests
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1
tenacity==8.2.3

# Data Processing & Analytics
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.2

# Time Series & ML
statsmodels==0.14.1
prophet==1.1.5
lightgbm==4.1.0

# Feature Engineering
feature-engine==1.6.2
category-encoders==2.6.3

# Data Validation & Quality
pydantic==2.5.2
great-expectations==0.18.5

# Task Scheduling & Workflow
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Apify API client for TikTok scraping
apify-client==1.7.1

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Vector Store & Embeddings (for semantic clustering)
chromadb==0.4.18
sentence-transformers==2.2.2

# Visualization & Dashboards
plotly==5.17.0
dash==2.16.1
streamlit==1.29.0

# Security & Authentication
cryptography==41.0.8
jwt==1.3.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
hashlib==20081119

# Optional: GPU acceleration for ML
# torch==2.1.2
# transformers==4.36.2 